+++ src/ui/components/Library/Team/View/Tests/Overview/TestOverviewContent.tsx
@@ -3,10 +3,10 @@ import { useContext } from "react";
 import { FailureRates } from "shared/test-suites/TestRun";
 import { LibrarySpinner } from "ui/components/Library/LibrarySpinner";
 
-import { useTest } from "../hooks/useTests";
 import { TestContext } from "../TestContextRoot";
 import { TestErrorList } from "./TestErrorList";
 import styles from "../../../../Library.module.css";
+import { useTest } from "../hooks/useTest";
 
 export function TestOverviewContent() {
   const { filterByText, testId, tests } = useContext(TestContext);
