+++ src/ui/components/Library/Team/View/Tests/graphql/TestsGraphQL.ts
@@ -26,35 +26,6 @@ const GET_TESTS = gql`
   }
 `;
 
-export const GET_TEST = gql`
-  query GetTestForWorkspace($workspaceId: ID!, $testId: String) {
-    node(id: $workspaceId) {
-      ... on Workspace {
-        id
-        tests(filter: { testId: $testId }) {
-          edges {
-            node {
-              testId
-              title
-              scope
-              executions {
-                errors
-                createdAt
-                commitTitle
-                result
-                recording {
-                  uuid
-                  title
-                }
-              }
-            }
-          }
-        }
-      }
-    }
-  }
-`;
-
 export async function getTestsGraphQL(
   graphQLClient: GraphQLClientInterface,
   accessToken: string | null,
